+++ R/utils-schema.R
@@ -1,6 +1,7 @@
 #' Get the JSON schema download URL for a given config file version
 #'
-#' @param config Name of config file to validate. One of `"tasks"` or `"admin"`.
+#' @param config Name of config file to validate. One of `"tasks"`, `"admin"`,
+#' `"model"` or `"target-data"`.
 #' @param version A valid version of hubverse
 #'   [schema](https://github.com/hubverse-org/schemas)
 #'   (e.g. `"v0.0.1"`).
@@ -14,8 +15,11 @@
 #'
 #' @examplesIf asNamespace("hubUtils")$not_rcmd_check()
 #' get_schema_url(config = "tasks", version = "v0.0.0.9")
-get_schema_url <- function(config = c("tasks", "admin", "model"),
-                           version, branch = "main") {
+get_schema_url <- function(
+  config = c("tasks", "admin", "model", "target-data"),
+  version,
+  branch = "main"
+) {
   config <- rlang::arg_match(config)
   rlang::check_required(version)
 
@@ -24,7 +28,9 @@ get_schema_url <- function(config = c("tasks", "admin", "model"),
   validate_schema_version(version, branch = branch)
 
   schema_repo <- "hubverse-org/schemas"
-  glue::glue("https://raw.githubusercontent.com/{schema_repo}/{branch}/{version}/{config}-schema.json")
+  glue::glue(
+    "https://raw.githubusercontent.com/{schema_repo}/{branch}/{version}/{config}-schema.json"
+  )
 }
 
 #' Get a vector of valid schema version
@@ -56,7 +62,8 @@ get_schema_valid_versions <- function(branch = "main") {
     ))
   }
 
-  req <- gh("GET /repos/hubverse-org/schemas/git/trees/{branch}",
+  req <- gh(
+    "GET /repos/hubverse-org/schemas/git/trees/{branch}",
     branch = branch
   )
 
@@ -89,9 +96,11 @@ get_schema <- function(schema_url) {
     if (fs::file_exists(path)) {
       return(jsonlite::prettify(readLines(path)))
     } else {
-      cli::cli_alert_warning("{.file {version}/{config}} not found.
+      cli::cli_alert_warning(
+        "{.file {version}/{config}} not found.
         This could mean your version of hubUtils is outdated.
-        Attempting to connect to GitHub.")
+        Attempting to connect to GitHub."
+      )
     }
   }
   response <- try(curl_fetch_memory(schema_url), silent = TRUE)
@@ -132,10 +141,10 @@ get_schema <- function(schema_url) {
 #' extract_schema_info(urls)
 extract_schema_info <- function(id) {
   lead <- "^https[:][/][/]raw.githubusercontent.com[/]hubverse-org[/]schemas[/]"
-  good_stuff <- "(.+?)[/](v[0-9.]+?)[/]([a-z]+?-schema.json)$"
+  good_stuff <- "(.+?)[/](v[0-9.]+?)[/]([a-z-]+?-schema.json)$"
   pattern <- paste0(lead, good_stuff)
-  proto <- setNames(character(3), c("branch", "version", "config"))
-  utils::strcapture(pattern, id, proto)
+  template <- setNames(character(3), c("branch", "version", "config"))
+  utils::strcapture(pattern, id, template)
 }
 
 #' Get the latest schema version
@@ -154,8 +163,10 @@ extract_schema_info <- function(id) {
 #' @examplesIf asNamespace("hubUtils")$not_rcmd_check()
 #' get_schema_version_latest()
 #' get_schema_version_latest(schema_version = "v3.0.0")
-get_schema_version_latest <- function(schema_version = "latest",
-                                      branch = "main") {
+get_schema_version_latest <- function(
+  schema_version = "latest",
+  branch = "main"
+) {
   if (schema_version == "latest") {
     get_schema_valid_versions(branch = branch) %>%
       sort() %>%
@@ -191,6 +202,7 @@ validate_schema_version <- function(schema_version, branch) {
 #' extract_schema_version("refs/heads/main/v3.0.0")
 extract_schema_version <- function(id) {
   stringr::str_extract(
-    id, "v[0-9]+\\.[0-9]+\\.[0-9]+(\\.9([0-9]+)?)?"
+    id,
+    "v[0-9]+\\.[0-9]+\\.[0-9]+(\\.9([0-9]+)?)?"
   )
 }